Improved robustness of non-interactive Wavelength solution
I have added several changes to Wavelength.py to improve the robustness of the non-interactive method.

Most of all, I have added a routine that uses sigma clipping to iteratively reject lines to achieve a pre-defined rms level.
It's a simple algorithm and probably can be improved but it seems to work well.
I also added an automatic fitting when a new slit is selected during the interactive fitting. Before, users had to manually hit "f"