Automatically measure the threshold to use to measure slit profiles
The threshold to measure slit profiles was set at 450. The user was able to change it in Driver.py by specifying it in the call to Flats.

To make the pipeline ready for KOA, I have added a routine to automatically determine the correct threshold. The routine measures the pixel values around the center of each slit, reduces that value to 80% of the measured results, and uses it as threshold.

I found that this solves the problem for K band observations, where the threshold sometimes has to be reduced to as little as 80 counts.